Support Ogre 14.5.1

+++ rviz_rendering/src/rviz_rendering/objects/point_cloud.cpp
@@ -649,7 +649,7 @@ size_t PointCloud::removePointsFromRenderables(
 
     size_t popped_in_renderable = std::min(
       static_cast<size_t>(number_of_points * vertices_per_point - popped_count),
-      op->vertexData->vertexCount);
+      static_cast<size_t>(op->vertexData->vertexCount));
     op->vertexData->vertexStart += popped_in_renderable;
     op->vertexData->vertexCount -= popped_in_renderable;
